Fixed issue for v4.0.3
Users that are not logged in or disabled sync had their local keyring erased on browser restart. It is now checked, if the user is logged in, before the sync on browser start is initiated. To test if the user is logged in, new "identity" permissions are required - sadly, the ID of the account requires acceess to the email as well. Other approaches with a minimal amount of permissions are welcome.